# swagger.yaml
openapi: 3.0.3
info:
  title: Smart Home Platform API
  description: REST API микросервисов экосистемы умного дома
  version: 1.0.0
servers:
  - url: https://api.smart-home.local
...
paths:
  /api/v1/devices/{deviceId}:
    get:
      tags: [Devices]
      summary: Получить информацию об устройстве
      parameters:
        - name: deviceId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Устройство найдено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Device'
        '404':
          description: Устройство не найдено

    patch:
      tags: [Devices]
      summary: Обновить логическое состояние устройства
      parameters:
        - name: deviceId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeviceUpdateRequest'
      responses:
        '200':
          description: Устройство обновлено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Device'
        '404':
          description: Устройство не найдено
...
components:
  schemas:
    Device:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        deviceType:
          type: string
          enum: [THERMOSTAT, LIGHT, CAMERA, GATE]
        houseId:
          type: string
        status:
          type: string
          enum: [ONLINE, OFFLINE]
        enabled:
          type: boolean
    DeviceUpdateRequest:
      type: object
      properties:
        enabled:
          type: boolean
...